♻️ refactor: InputBox common 폴더로 이동
