+++ src/components/common/InputBox/InputBox.stories.tsx
