+++ src/components/create/index.ts
@@ -1,7 +1,7 @@
 import SelectBox from "./SelectBox";
 import PartBox from "./PartBox";
 import Select from "./Select";
-import InputBox from "./InputBox/InputBox";
+import InputBox from "../common/InputBox/InputBox";
 import ImageUploader from "./ImageUploader/ImageUploader";
 import PartBoxList from "./PartBoxList";
 import DatePicker from "./DatePicker";
